requirements.txt
python-dotenv~=1.0.1
mistralai~=1.5.0
pymongo~=4.11
langchain~=0.3.20
langchain-core~=0.3.45